vmm: uintptr_t for the frame-map pageAddr field (cross-arch Phase 7, 2/n)
vmm_page_info_t.pageAddr holds a physical frame address — retype it uintptr_t
and widen the i*PAGE_SIZE init so frame addresses don't truncate above 4 GB on a
64-bit target.  Codegen-neutral on i386 (uintptr_t == unsigned int).

+++ sys/include/vmm/vmm.h
@@ -52,7 +52,7 @@ extern "C" {
     extern struct spinLock pdSpinLock;
 
     typedef struct {
-            u_int32_t pageAddr;
+            uintptr_t pageAddr;
             u_int16_t status;
             u_int16_t reserved;
             pid_t pid;
